requirements.txt
httpx>=0.27
jinja2>=3.1
pyyaml>=6.0
